backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.1
gunicorn==23.0.0
joblib==1.5.0
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
scikit-learn==1.6.1
